% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LdFlags.R
\name{LdFlags}
\alias{LdFlags}
\title{Get portable linker flags for libraries building on RcppThread}
\usage{
LdFlags()
}
\description{
To be used in \code{Makevars} on Linux and OSX. Returns a character vector with
linker flags for \code{pthread} and \code{libatomic}, if available.
}
\details{
Use as
`PKG_LIBS = ${R_HOME}/bin/Rscript -e 'RcppThread::LdFlags()'.
}
