% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppThread-package.R
\docType{package}
\name{RcppThread}
\alias{RcppThread}
\alias{RcppThread-package}
\title{R-friendly C++11 threads}
\description{
Provides a C++11-style thread class and thread pool that can safely be
interrupted from R.
}
