% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppThreads-package.r
\docType{package}
\name{RcppThreads}
\alias{RcppThreads}
\alias{RcppThreads-package}
\title{R-friendly C++11 threads}
\description{
Provides a C++11-style thread class and thread pool that can safely be
interrupted from R.
}
